ref(replay): Remove FluidPanel abstraction
